minor .gitignore tweak
in some environments, python won't be able to create the directory it needs for css. so I include it with a gitignore file inside of it that says to ignore all the css files that are compiled from sass. git won't include empty folders, so this file is essential and I don't want the css files on my repo at all.